+++ packages/sdks/core-js-sdk/src/sdk/types.ts
@@ -346,6 +346,7 @@ export type Options = {
   locale?: string;
   oidcPrompt?: string;
   oidcErrorRedirectUri?: string;
+  oidcResource?: string;
   nativeOptions?: NativeOptions;
   thirdPartyAppStateId?: string;
   applicationScopes?: string; // Relevant for sso application and third party application