+++ packages/sdks/web-component/test/descope-wc.test.ts
@@ -30,6 +30,7 @@ import {
   OIDC_PROMPT_PARAM_NAME,
   SDK_SCRIPT_RESULTS_KEY,
   OIDC_ERROR_REDIRECT_URI_PARAM_NAME,
+  OIDC_RESOURCE_PARAM_NAME,
   THIRD_PARTY_APP_STATE_ID_PARAM_NAME,
   APPLICATION_SCOPES_PARAM_NAME,
   SDK_SCRIPTS_LOAD_TIMEOUT,
@@ -52,7 +53,7 @@ jest.mock('@descope/web-js-sdk', () => ({
   ensureFingerprintIds: jest.fn(),
 }));
 
-const WAIT_TIMEOUT = 20000;
+const WAIT_TIMEOUT = 25000;
 const THEME_DEFAULT_FILENAME = `theme.json`;
 
 const abTestingKey = getABTestingKey();
@@ -68,6 +69,7 @@ const defaultOptionsValues = {
   samlIdpStateId: null,
   samlIdpUsername: null,
   oidcErrorRedirectUri: null,
+  oidcResource: null,
   descopeIdpInitiated: false,
   ssoAppId: null,
   client: {},
@@ -385,18 +387,14 @@ describe('web-component', () => {
 
     await waitFor(
       () => {
-        expect(screen.getByShadowText('Another Button')).not.toHaveAttribute(
-          'disabled',
-        );
+        expect(screen.getByShadowText('Another Button')).toBeEnabled();
       },
       { timeout: WAIT_TIMEOUT },
     );
 
     await waitFor(
       () => {
-        expect(screen.getByShadowPlaceholderText('Input')).not.toHaveAttribute(
-          'disabled',
-        );
+        expect(screen.getByShadowPlaceholderText('Input')).toBeEnabled();
       },
       { timeout: WAIT_TIMEOUT },
     );
@@ -456,14 +454,14 @@ describe('web-component', () => {
 
     await waitFor(
       () => {
-        expect(anotherButton).not.toHaveAttribute('disabled');
+        expect(anotherButton).toBeEnabled();
       },
       { timeout: WAIT_TIMEOUT },
     );
 
     await waitFor(
       () => {
-        expect(inputField).not.toHaveAttribute('disabled');
+        expect(inputField).toBeEnabled();
       },
       { timeout: WAIT_TIMEOUT },
     );
@@ -3920,6 +3918,74 @@ describe('web-component', () => {
     const encodedOidcErrorRedirectUri =
       encodeURIComponent(oidcErrorRedirectUri);
     window.location.search = `?${OIDC_ERROR_REDIRECT_URI_PARAM_NAME}=${encodedOidcErrorRedirectUri}`;
+    document.body.innerHTML = `<h1>Custom element test</h1> <descope-wc flow-id="sign-in" project-id="1"></descope-wc>`;
+
+    await waitFor(() => expect(startMock).toHaveBeenCalled(), {
+      timeout: WAIT_TIMEOUT,
+    });
+
+    await waitFor(() => screen.getByShadowText('It works!'), {
+      timeout: WAIT_TIMEOUT,
+    });
+
+    fireEvent.click(screen.getByShadowText('click'));
+
+    await waitFor(() => expect(nextMock).toHaveBeenCalled());
+  });
+
+  it('should call start with oidc idp with oidcResource flag and clear it from url', async () => {
+    startMock.mockReturnValueOnce(generateSdkResponse());
+
+    pageContent = '<span>It works!</span>';
+
+    const oidcStateId = 'abcdefgh';
+    const encodedOidcStateId = encodeURIComponent(oidcStateId);
+    const oidcResource = 'https://api.example.com';
+    const encodedOidcResource = encodeURIComponent(oidcResource);
+    window.location.search = `?${OIDC_IDP_STATE_ID_PARAM_NAME}=${encodedOidcStateId}&${OIDC_RESOURCE_PARAM_NAME}=${encodedOidcResource}`;
+    document.body.innerHTML = `<h1>Custom element test</h1> <descope-wc flow-id="sign-in" project-id="1"></descope-wc>`;
+
+    await waitFor(
+      () =>
+        expect(startMock).toHaveBeenCalledWith(
+          'sign-in',
+          {
+            ...defaultOptionsValues,
+            oidcIdpStateId: 'abcdefgh',
+            oidcResource: 'https://api.example.com',
+          },
+          undefined,
+          '',
+          '1.2.3',
+          {
+            otpSignInEmail: 1,
+            'versioned-flow': 1,
+          },
+          {},
+        ),
+      { timeout: WAIT_TIMEOUT },
+    );
+    await waitFor(() => screen.getByShadowText('It works!'), {
+      timeout: WAIT_TIMEOUT,
+    });
+    await waitFor(() => expect(window.location.search).toBe(''));
+  });
+
+  it('should call start with oidc idp with oidcResource when there is a start screen is configured', async () => {
+    startMock.mockReturnValueOnce(generateSdkResponse());
+
+    configContent = {
+      flows: {
+        'sign-in': { startScreenId: 'screen-0' },
+      },
+    };
+
+    pageContent =
+      '<descope-button>click</descope-button><span>It works!</span>';
+
+    const oidcResource = 'https://api.example.com';
+    const encodedOidcResource = encodeURIComponent(oidcResource);
+    window.location.search = `?${OIDC_RESOURCE_PARAM_NAME}=${encodedOidcResource}`;
     document.body.innerHTML = `<h1>Custom element test</h1> <descope-wc flow-id="sign-in" project-id="1"></descope-wc>`;
 
     await waitFor(() => expect(startMock).toHaveBeenCalled(), {