+++ packages/sdks/vue-sdk/CHANGELOG.md
@@ -2,6 +2,38 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [2.10.9](https://github.com/descope/descope-js/compare/vue-sdk-2.10.8...vue-sdk-2.10.9) (2025-08-14)
+
+### Dependency Updates
+
+* `access-key-management-widget` updated to version `0.5.3`
+* `audit-management-widget` updated to version `0.5.3`
+* `role-management-widget` updated to version `0.4.3`
+* `user-management-widget` updated to version `0.9.0`
+* `user-profile-widget` updated to version `0.6.8`
+* `applications-portal-widget` updated to version `0.4.3`
+* `web-component` updated to version `3.46.0`
+* `web-js-sdk` updated to version `1.34.0`
+* `core-js-sdk` updated to version `2.46.0`
+## [2.10.8](https://github.com/descope/descope-js/compare/vue-sdk-2.10.7...vue-sdk-2.10.8) (2025-08-10)
+
+### Dependency Updates
+
+* `user-profile-widget` updated to version `0.6.7`
+* `web-component` updated to version `3.45.1`
+## [2.10.7](https://github.com/descope/descope-js/compare/vue-sdk-2.10.6...vue-sdk-2.10.7) (2025-08-07)
+
+### Dependency Updates
+
+* `access-key-management-widget` updated to version `0.5.2`
+* `audit-management-widget` updated to version `0.5.2`
+* `role-management-widget` updated to version `0.4.2`
+* `user-management-widget` updated to version `0.8.2`
+* `user-profile-widget` updated to version `0.6.6`
+* `applications-portal-widget` updated to version `0.4.2`
+* `web-component` updated to version `3.45.0`
+* `web-js-sdk` updated to version `1.33.7`
+* `core-js-sdk` updated to version `2.45.0`
 ## [2.10.6](https://github.com/descope/descope-js/compare/vue-sdk-2.10.5...vue-sdk-2.10.6) (2025-08-05)
 
 ### Dependency Updates