+++ packages/widgets/tenant-profile-widget/CHANGELOG.md
@@ -2,6 +2,28 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.2.5](https://github.com/descope/descope-js/compare/tenant-profile-widget-0.2.4...tenant-profile-widget-0.2.5) (2025-08-14)
+
+### Dependency Updates
+
+* `core-js-sdk` updated to version `2.46.0`
+* `web-js-sdk` updated to version `1.34.0`
+* `web-component` updated to version `3.46.0`
+* `sdk-helpers` updated to version `0.4.0`
+* `sdk-mixins` updated to version `0.13.10`
+* `sdk-component-drivers` updated to version `0.6.0`
+## [0.2.4](https://github.com/descope/descope-js/compare/tenant-profile-widget-0.2.3...tenant-profile-widget-0.2.4) (2025-08-10)
+
+### Dependency Updates
+
+* `web-component` updated to version `3.45.1`
+## [0.2.3](https://github.com/descope/descope-js/compare/tenant-profile-widget-0.2.2...tenant-profile-widget-0.2.3) (2025-08-07)
+
+### Dependency Updates
+
+* `core-js-sdk` updated to version `2.45.0`
+* `web-js-sdk` updated to version `1.33.7`
+* `web-component` updated to version `3.45.0`
 ## [0.2.2](https://github.com/descope/descope-js/compare/tenant-profile-widget-0.2.1...tenant-profile-widget-0.2.2) (2025-08-05)
 
 ### Dependency Updates