+++ packages/widgets/user-management-widget/CHANGELOG.md
@@ -2,6 +2,25 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.9.0](https://github.com/descope/descope-js/compare/user-management-widget-0.8.2...user-management-widget-0.9.0) (2025-08-14)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.34.0`
+* `web-component` updated to version `3.46.0`
+* `sdk-helpers` updated to version `0.4.0`
+* `sdk-mixins` updated to version `0.13.10`
+* `sdk-component-drivers` updated to version `0.6.0`
+
+### Features
+
+* Generic flow button ([#1172](https://github.com/descope/descope-js/issues/1172)) ([9ac9e8c](https://github.com/descope/descope-js/commit/9ac9e8c7fe34fce0d8bd26ec7a824d902a8208ec))
+
+## [0.8.2](https://github.com/descope/descope-js/compare/user-management-widget-0.8.1...user-management-widget-0.8.2) (2025-08-07)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.33.7`
 ## [0.8.1](https://github.com/descope/descope-js/compare/user-management-widget-0.8.0...user-management-widget-0.8.1) (2025-08-05)
 
 ### Dependency Updates